Major Restructure and addition of unit test class for decks and unit testing project generally
