+++ MLBlackjack.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29215.179
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLBlackjack", "MLBlackjack.csproj", "{A70C5198-F79F-444E-BE57-672DC015E59B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MLBlackjack", "MLBlackjack\MLBlackjack.csproj", "{A83C91C4-543D-4275-8BBA-8EA367285674}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MLBlackjack.Testing", "..\MLBlackjack.Testing\MLBlackjack.Testing.csproj", "{1C87FBEB-C98A-4C98-9308-D4A22EBE92A0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A70C5198-F79F-444E-BE57-672DC015E59B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A70C5198-F79F-444E-BE57-672DC015E59B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A70C5198-F79F-444E-BE57-672DC015E59B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A70C5198-F79F-444E-BE57-672DC015E59B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A83C91C4-543D-4275-8BBA-8EA367285674}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A83C91C4-543D-4275-8BBA-8EA367285674}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A83C91C4-543D-4275-8BBA-8EA367285674}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A83C91C4-543D-4275-8BBA-8EA367285674}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C87FBEB-C98A-4C98-9308-D4A22EBE92A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C87FBEB-C98A-4C98-9308-D4A22EBE92A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C87FBEB-C98A-4C98-9308-D4A22EBE92A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C87FBEB-C98A-4C98-9308-D4A22EBE92A0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
