+++ MLBlackjack/models/PlayerAction.cs
@@ -1,0 +1,8 @@
+namespace CardExploration.models
+{
+    public enum PlayerAction
+    {
+        stand,
+        hit
+    }
+}
