+++ MLBlackjack/models/phase.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace CardExploration.models
+{
+    //phase of player in game
+    public enum Phase
+    {
+        bust,
+        revealCards,
+        stand
+    }
+}
