--- models/PlayerAction.cs
@@ -1,8 +1,0 @@
-namespace CardExploration.models
-{
-    public enum PlayerAction
-    {
-        stand,
-        hit
-    }
-}
