--- models/phase.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace CardExploration.models
-{
-    //phase of player in game
-    public enum Phase
-    {
-        bust,
-        revealCards,
-        stand
-    }
-}
